test: move PR fetch integration tests out of the hook-run suite
Revert the production git env scrub (candidates.ts, branch-fallback.ts):
production git spawns inherit the environment again, keeping the plugin
minimal. Bun snapshots env at process start for children, so a test-side
env strip cannot shield production runGit calls from GIT_DIR; instead the
two real-git sandbox tests now live in candidates.integration.test.ts,
excluded from `bun test` (--path-ignore-patterns) and run only via
`bun run test:integration`. CI runs both suites; the husky pre-commit
hook runs the unit suite only, so no sandbox git ever executes under the
hook's GIT_DIR environment. The integration suite keeps its unique
tmpdir sandbox and explicit-env git helper.

Refs #33
