+++ .github/workflows/ci.yml
@@ -23,4 +23,6 @@ jobs:
         run: bun run typecheck
 
       - name: Test
-        run: bun test
+        run: |
+          bun run test
+          bun run test:integration
